refactor : library 통합
